Added is_active check to vehicles model
Added an is_active boolean column to the vehicles table to help track past and present vehicles in the Kingdomhire fleet. This was done as we needed to keep track of all hires (past and present), and if a vehicle was removed from the DB, we will run into issues with foreign key relations between the vehicles and hires tables, as the vehicle assosciated with a hire would no longer exist. So the solution for now is to re-do the process of deleting a vehicle into more of a vehicle discontinuing process, were the vehicle will still persist in the vehicles table so we can still generate reports on these vehicles in the future with out any issue