Added reservation/hire edit features
You can now re-schedule a reservation and shorten/extend a hire. These features are different as with reservations you can can both the start and end dates, whereas with hires you can only change the end date. Will need to add validation on the backend to compare the start and end dates to make sure they are correct