Minor Feature - Image to Vehicle link
Now each vehicle can have more than 1 image associated with it. Each vehicle will have it's own images folder to store all the images linked to that vehicle. Will need to use javascript to allow the user to look through all the images associated with a particular vehicle in a gallery format.