Updated validation errors/rules across most forms
Some forms didn't display error messages the same way as other, all forms are now consistent. Also separated error messages for editing a vehicle and adding/editing a reservation into seperate error bags, as these forms are all displayed in the vehicle dashboard, just in different tabs. So now when errors are returned to the view, the correct tab will be displayed that is associated with those error messages