Implemented session status messages + re-added GUI for delete forms
Session status messages will now be properly displayed when a form is submitted, such as when you book a reservation, you will be notified that it was successfully in an alert message that will be displayed on the screen. Also re-added GUI components for deleting/discontining a vehicle and for deleting a weekly rate. Some JavaScript will be needed to fade out the session status messages, as they currently persist until the page is refreshed