Vehicle route keys + model id changes
Now vehicle routes will contain a slug instead of just the vehicle id e.g /admin/vehicles/kia-sedona-1467. Also vehicle ids are now a 4 digit number stored as a string in the database as the primary key instead of an auto incrementing integer as the primary key. These id will be used as a way to more uniquely identify a vehicle